pyproject.toml
[project]
name = "diagnosis-asst"
version = "0.1.0"
description = ""
authors = [
    {name = "Pradeepaa001",email = "mpradeepaa2020@gmail.com"}
]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10,<4"
torch = "*"
torchvision = "*"
tensorflow = "*"
numpy = "*"
pandas = "*"
scikit-learn = "*"
flask = "*"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
opencv-python = "*"
Pillow = "*"
streamlit = "*"
google-generativeai = "*"
federated-learning = "*"
python-dotenv = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"